requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.6
djangorestframework==3.16.0
filelock==3.17.0
mysqlclient==2.2.7
numpy==2.2.5
pandas==2.2.3
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
virtualenv==20.29.2
